Extract transfer instructions to instructions_transfer.go module
